+++ odradek-app/src/main/java/sh/adelessfox/odradek/app/ui/tools/graph/GraphToolPanel.java
@@ -21,12 +21,12 @@ import sh.adelessfox.odradek.ui.Focusable;
 import sh.adelessfox.odradek.ui.actions.Actions;
 import sh.adelessfox.odradek.ui.components.SearchTextField;
 import sh.adelessfox.odradek.ui.components.ValidationPopup;
-import sh.adelessfox.odradek.ui.components.tools.ToolPanel;
-import sh.adelessfox.odradek.ui.components.tools.ToolSite;
 import sh.adelessfox.odradek.ui.components.tree.StructuredTree;
 import sh.adelessfox.odradek.ui.components.tree.TreeActionListener;
 import sh.adelessfox.odradek.ui.components.tree.TreeLabelProvider;
 import sh.adelessfox.odradek.ui.data.DataKeys;
+import sh.adelessfox.odradek.ui.tools.ToolPanel;
+import sh.adelessfox.odradek.ui.tools.ToolSite;
 import sh.adelessfox.odradek.ui.util.Fugue;
 import sh.adelessfox.odradek.util.Result;
