--- odradek-ui/src/main/java/sh/adelessfox/odradek/ui/components/Orientation.java
@@ -1,6 +1,0 @@
-package sh.adelessfox.odradek.ui.components;
-
-public enum Orientation {
-    HORIZONTAL,
-    VERTICAL
-}
